+++ includes/features/class-wpcli.php
@@ -23,6 +23,8 @@ use Traffic\System\Option;
 use Traffic\System\GeoIP;
 use Traffic\System\Timezone;
 use Traffic\System\UUID;
+use Traffic\System\Http;
+use Traffic\System\SharedMemory;
 
 /**
  * WP-CLI for Traffic.
@@ -300,64 +302,39 @@ class Wpcli {
 	}
 
 	/**
-	 * Get DecaLog details and operation modes.
+	 * Get Traffic details and operation modes.
 	 *
 	 * ## EXAMPLES
 	 *
-	 * wp log status
+	 * wp api status
 	 *
 	 *
 	 *   === For other examples and recipes, visit https://github.com/Pierre-Lannoy/wp-traffic/blob/master/WP-CLI.md ===
 	 *
 	 */
 	public static function status( $args, $assoc_args ) {
-		$run  = 0;
-		$list = 0;
-		foreach ( Option::network_get( 'loggers' ) as $key => $logger ) {
-			if ( $logger['running'] ) {
-				$run++;
-			}
-		}
-		if ( Option::network_get( 'autolisteners' ) ) {
-			$list = 'on all available listeners';
+		\WP_CLI::line( sprintf( '%s is running.', Environment::plugin_version_text() ) );
+		if ( Option::network_get( 'inbound_capture' ) ) {
+			\WP_CLI::line( 'Inbound analytics: enabled.' );
 		} else {
-			$listeners = ListenerFactory::$infos;
-			foreach ( $listeners as $listener ) {
-				if ( $listener['available'] && in_array( $listener['id'], Option::network_get( 'listeners' ), true ) ) {
-					$list++;
-				}
-			}
-			if ( 0 === $list) {
-				$list = 'on no listener';
-			} elseif ( 1 === $list) {
-				$list = 'on 1 listener';
-			} else {
-				$list = sprintf( 'on %d listeners', $list );
-			}
+			\WP_CLI::line( 'Inbound analytics: disabled.' );
 		}
-		if ( 0 === $run) {
-			$run  = '';
-			$list = '';
-		} elseif ( 1 === $run) {
-			$run = '1 logger';
+		\WP_CLI::line( 'Inbound logging: ' . Option::network_get( 'inbound_level' ) . '.' );
+		if ( Option::network_get( 'outbound_capture' ) ) {
+			\WP_CLI::line( 'Outbound analytics: enabled.' );
 		} else {
-			$run = sprintf( '%d loggers', $run );
+			\WP_CLI::line( 'Outbound analytics: disabled.' );
 		}
-		\WP_CLI::line( sprintf( '%s running %s %s.', Environment::plugin_version_text(), $run, $list ) );
-		if ( Option::network_get( 'earlyloading' ) ) {
-			\WP_CLI::line( 'Early-Loading: enabled.' );
+		\WP_CLI::line( 'Outbound logging: ' . Option::network_get( 'outbound_level' ) . '.' );
+		if ( Option::network_get( 'livelog' ) ) {
+			\WP_CLI::line( 'Auto-Monitoring: enabled.' );
 		} else {
-			\WP_CLI::line( 'Early-Loading: disabled.' );
+			\WP_CLI::line( 'Auto-Monitoring: disabled.' );
 		}
-		if ( Option::network_get( 'logger_autostart' ) ) {
-			\WP_CLI::line( 'Auto-Start: enabled.' );
+		if ( defined( 'DECALOG_VERSION' ) ) {
+			\WP_CLI::line( 'Logging support: yes (DecaLog v' . DECALOG_VERSION . ').');
 		} else {
-			\WP_CLI::line( 'Auto-Start: disabled.' );
-		}
-		if ( Autolog::is_enabled() ) {
-			\WP_CLI::line( 'Auto-Logging: enabled.' );
-		} else {
-			\WP_CLI::line( 'Auto-Logging: disabled.' );
+			\WP_CLI::line( 'Logging support: no.' );
 		}
 		$geo = new GeoIP();
 		if ( $geo->is_installed() ) {
@@ -365,616 +342,11 @@ class Wpcli {
 		} else {
 			\WP_CLI::line( 'IP information support: no.' );
 		}
-		if ( class_exists( 'PODeviceDetector\API\Device' ) ) {
-			\WP_CLI::line( 'Device detection support: yes (Device Detector v' . PODD_VERSION . ').');
+		if ( SharedMemory::$available ) {
+			\WP_CLI::line( 'Shared memory support: yes (shmop v' . phpversion( 'shmop' ) . ').');
 		} else {
-			\WP_CLI::line( 'Device detection support: no.' );
+			\WP_CLI::line( 'Shared memory support: no.' );
 		}
-	}
-
-	/**
-	 * Get information on logger types.
-	 *
-	 * ## OPTIONS
-	 *
-	 * <list|describe>
-	 * : The action to take.
-	 * ---
-	 * options:
-	 *  - list
-	 *  - describe
-	 * ---
-	 *
-	 * [<logger_type>]
-	 * : The type of the logger to describe.
-	 *
-	 * [--format=<format>]
-	 * : Allows overriding the output of the command when listing types.
-	 * ---
-	 * default: table
-	 * options:
-	 *  - table
-	 *  - json
-	 *  - csv
-	 *  - yaml
-	 *  - ids
-	 *  - count
-	 * ---
-	 *
-	 * ## EXAMPLES
-	 *
-	 * Lists available types:
-	 * + wp log type list
-	 * + wp log type list --format=json
-	 *
-	 * Details the WordpressHandler logger type:
-	 * + wp log type describe WordpressHandler
-	 *
-	 *
-	 *   === For other examples and recipes, visit https://github.com/Pierre-Lannoy/wp-traffic/blob/master/WP-CLI.md ===
-	 *
-	 */
-	public static function handler( $args, $assoc_args ) {
-		$stdout        = \WP_CLI\Utils\get_flag_value( $assoc_args, 'stdout', false );
-		$format        = \WP_CLI\Utils\get_flag_value( $assoc_args, 'format', 'table' );
-		$handler_types = new HandlerTypes();
-		$handlers      = [];
-		foreach ( $handler_types->get_all() as $key => $handler ) {
-			if ( 'system' !== $handler['class'] ) {
-				$handler['type']                        = $handler['id'];
-				$handlers[strtolower( $handler['id'] )] = $handler;
-			}
-		}
-		uasort(
-			$handlers,
-			function ( $a, $b ) {
-				return strcmp( strtolower( $a[ 'name' ] ), strtolower( $b[ 'name' ] ) );
-			}
-		);
-		$uuid   = '';
-		$action = isset( $args[0] ) ? $args[0] : 'list';
-		if ( isset( $args[1] ) ) {
-			$uuid = strtolower( $args[1] );
-			if ( ! array_key_exists( $uuid, $handlers ) ) {
-				$uuid = '';
-			}
-		}
-		if ( 'list' !== $action && '' === $uuid ) {
-			self::error( 1, $stdout );
-		}
-		switch ( $action ) {
-			case 'list':
-				if ( 'ids' === $format ) {
-					self::write_ids( $handlers, 'type' );
-				} else {
-					\WP_CLI\Utils\format_items( $format, $handlers, [ 'type', 'class', 'name', 'version' ] );
-				}
-				break;
-			case 'describe':
-				$example = [];
-				$handler = $handlers[$uuid];
-				\WP_CLI::line( '' );
-				\WP_CLI::line( \WP_CLI::colorize( '%8' . $handler['name'] . ' - ' . $handler['id'] . '%n' ) );
-				\WP_CLI::line( $handler['help'] );
-				\WP_CLI::line( '' );
-				\WP_CLI::line( \WP_CLI::colorize( '%UMinimal Level%n' ) );
-				\WP_CLI::line( '' );
-				\WP_CLI::line( '  ' .  strtolower( Log::level_name( $handler['minimal'] ) ) );
-				\WP_CLI::line( '' );
-				\WP_CLI::line( \WP_CLI::colorize( '%UParameters%n' ) );
-				\WP_CLI::line( '' );
-				$param = '  * ';
-				$elem  = '    - ';
-				$list  = '       ';
-				\WP_CLI::line( $param . 'Name - Used only in admin dashboard.' );
-				\WP_CLI::line( $elem . 'field name: name' );
-				\WP_CLI::line( $elem . 'field type: string' );
-				\WP_CLI::line( $elem . 'default value: "New Logger"' );
-				\WP_CLI::line( '' );
-				\WP_CLI::line( $param . 'Minimal level - Minimal reported level.' );
-				\WP_CLI::line( $elem . 'field name: level' );
-				\WP_CLI::line( $elem . 'field type: string' );
-				\WP_CLI::line( $elem . 'default value: "' . strtolower( Log::level_name( $handler['minimal'] ) ) . '"' );
-				\WP_CLI::line( $elem . 'available values:' );
-				foreach ( Log::get_levels( EventTypes::$levels[ strtolower( Log::level_name( $handler['minimal'] ) ) ] ) as $level ) {
-					\WP_CLI::line( $list . '"' . strtolower( $level[1] ) . '": ' . $level[2]);
-				}
-				\WP_CLI::line( '' );
-				foreach ( $handler['configuration'] as $key => $conf ) {
-					if ( ! $conf['show'] || ! $conf['control']['enabled'] ) {
-						continue;
-					}
-					\WP_CLI::line( $param . $conf['name'] . ' - ' . $conf['help'] );
-					\WP_CLI::line( $elem . 'field name: ' . $key );
-					\WP_CLI::line( $elem . 'field type: ' . $conf['type'] );
-					switch ( $conf['control']['type'] ) {
-						case 'field_input_integer':
-							\WP_CLI::line( $elem . 'default value: ' . $conf['default'] );
-							\WP_CLI::line( $elem . 'range: [' . $conf['control']['min'] . '-' . $conf['control']['max'] . ']' );
-							$example[] = '"' . $key . '": ' . $conf['default'];
-							break;
-						case 'field_checkbox':
-							\WP_CLI::line( $elem . 'default value: ' . ( $conf['default'] ? 'true' : 'false' ) );
-							$example[] = '"' . $key . '": ' . ( $conf['default'] ? 'true' : 'false' );
-							break;
-						case 'field_input_text':
-							\WP_CLI::line( $elem . 'default value: "' . $conf['default'] . '"' );
-							$example[] = '"' . $key . '": "' . $conf['default'] . '"';
-							break;
-						case 'field_select':
-							switch ( $conf['control']['cast'] ) {
-								case 'integer':
-									\WP_CLI::line( $elem . 'default value: ' . $conf['default'] );
-									$example[] = '"' . $key . '": ' . $conf['default'];
-									break;
-								case 'string':
-									\WP_CLI::line( $elem . 'default value: "' . $conf['default'] . '"' );
-									$example[] = '"' . $key . '": "' . $conf['default'] . '"';
-									break;
-							}
-							\WP_CLI::line( $elem . 'available values:' );
-							foreach ( $conf['control']['list'] as $point ) {
-								switch ( $conf['control']['cast'] ) {
-									case 'integer':
-										\WP_CLI::line( $list . $point[0] . ': ' . $point[1]);
-										break;
-									case 'string':
-										\WP_CLI::line( $list . '"' . $point[0] . '": ' . $point[1]);
-										break;
-								}
-							}
-							break;
-					}
-					\WP_CLI::line( '' );
-				}
-				\WP_CLI::line( $param . 'IP obfuscation - Log fields will contain hashes instead of real IPs.' );
-				\WP_CLI::line( $elem . 'field name: obfuscation' );
-				\WP_CLI::line( $elem . 'field type: boolean' );
-				\WP_CLI::line( $elem . 'default value: false' );
-				\WP_CLI::line( '' );
-				\WP_CLI::line( $param . 'User pseudonymization - Log fields will contain hashes instead of user IDs & names.' );
-				\WP_CLI::line( $elem . 'field name: pseudonymization' );
-				\WP_CLI::line( $elem . 'field type: boolean' );
-				\WP_CLI::line( $elem . 'default value: false' );
-				\WP_CLI::line( '' );
-				\WP_CLI::line( $param . 'Reported details: WordPress - Allows to log site, user and remote IP of the current request.' );
-				\WP_CLI::line( $elem . 'field name: proc_wp' );
-				\WP_CLI::line( $elem . 'field type: boolean' );
-				\WP_CLI::line( $elem . 'default value: true' );
-				\WP_CLI::line( '' );
-				\WP_CLI::line( $param . 'Reported details: HTTP request - Allows to log url, method, referrer and remote IP of the current web request.' );
-				\WP_CLI::line( $elem . 'field name: proc_http' );
-				\WP_CLI::line( $elem . 'field type: boolean' );
-				\WP_CLI::line( $elem . 'default value: true' );
-				\WP_CLI::line( '' );
-				\WP_CLI::line( $param . 'Reported details: PHP introspection - Allows to log line, file, class and function generating the event.' );
-				\WP_CLI::line( $elem . 'field name: proc_php' );
-				\WP_CLI::line( $elem . 'field type: boolean' );
-				\WP_CLI::line( $elem . 'default value: true' );
-				\WP_CLI::line( '' );
-				\WP_CLI::line( $param . 'Reported details: Backtrace - Allows to log the full PHP and WordPress call stack.' );
-				\WP_CLI::line( $elem . 'field name: proc_trace' );
-				\WP_CLI::line( $elem . 'field type: boolean' );
-				\WP_CLI::line( $elem . 'default value: false' );
-				\WP_CLI::line( '' );
-				\WP_CLI::line( \WP_CLI::colorize( '%UExample%n' ) );
-				\WP_CLI::line( '' );
-				\WP_CLI::line( '  {' . implode( ', ', $example ) . '}' );
-				\WP_CLI::line( '' );
-				break;
-		}
-
-	}
-
-	/**
-	 * Manage Decalog loggers.
-	 *
-	 * ## OPTIONS
-	 *
-	 * <list|start|pause|clean|purge|remove|add|set>
-	 * : The action to take.
-	 * ---
-	 * options:
-	 *  - list
-	 *  - start
-	 *  - pause
-	 *  - clean
-	 *  - purge
-	 *  - remove
-	 *  - add
-	 *  - set
-	 * ---
-	 *
-	 * [<uuid_or_type>]
-	 * : The uuid of the logger to perform an action on or the type of the logger to add.
-	 *
-	 * [--settings=<settings>]
-	 * : The settings needed by "add" and "modify" actions.
-	 * MUST be a string containing a json configuration.
-	 * ---
-	 * default: '{}'
-	 * example: '{"host": "syslog.collection.eu.sumologic.com", "timeout": 800, "ident": "DecaLog", "format": 1}'
-	 * ---
-	 *
-	 * [--detail=<detail>]
-	 * : The details of the output when listing loggers.
-	 * ---
-	 * default: short
-	 * options:
-	 *  - short
-	 *  - full
-	 * ---
-	 *
-	 * [--format=<format>]
-	 * : Allows overriding the output of the command when listing loggers.
-	 * ---
-	 * default: table
-	 * options:
-	 *  - table
-	 *  - json
-	 *  - csv
-	 *  - yaml
-	 *  - ids
-	 *  - count
-	 * ---
-	 *
-	 * [--yes]
-	 * : Answer yes to the confirmation message, if any.
-	 *
-	 * [--stdout]
-	 * : Use clean STDOUT output to use results in scripts. Unnecessary when piping commands because piping is detected by DecaLog.
-	 *
-	 * ## EXAMPLES
-	 *
-	 * Lists configured loggers:
-	 * + wp log logger list
-	 * + wp log logger list --detail=full
-	 * + wp log logger list --format=json
-	 *
-	 * Starts a logger:
-	 * + wp log logger start 37cf1c00-d67d-4e7d-9518-e579f01407a7
-	 *
-	 * Pauses a logger:
-	 * + wp log logger pause 37cf1c00-d67d-4e7d-9518-e579f01407a7
-	 *
-	 * Deletes old records of a logger:
-	 * + wp log logger clean 37cf1c00-d67d-4e7d-9518-e579f01407a7
-	 *
-	 * Deletes all records of a logger:
-	 * + wp log logger purge 37cf1c00-d67d-4e7d-9518-e579f01407a7
-	 * + wp log logger purge 37cf1c00-d67d-4e7d-9518-e579f01407a7 --yes
-	 *
-	 * Permanently deletes a logger:
-	 * + wp log logger remove 37cf1c00-d67d-4e7d-9518-e579f01407a7
-	 * + wp log logger remove 37cf1c00-d67d-4e7d-9518-e579f01407a7 --yes
-	 *
-	 * Adds a new logger:
-	 * + wp log logger add WordpressHandler --settings='{"rotate": 8000, "purge": 5, "level":"warning", "proc_wp": true}'
-	 *
-	 * Change the settings of a logger
-	 * + wp log logger set 37cf1c00-d67d-4e7d-9518-e579f01407a7 --settings='{"proc_trace": false, "level":"warning"}'
-	 *
-	 *
-	 *   === For other examples and recipes, visit https://github.com/Pierre-Lannoy/wp-traffic/blob/master/WP-CLI.md ===
-	 *
-	 */
-	public static function logger( $args, $assoc_args ) {
-		$stdout       = \WP_CLI\Utils\get_flag_value( $assoc_args, 'stdout', false );
-		$format       = \WP_CLI\Utils\get_flag_value( $assoc_args, 'format', 'table' );
-		$detail       = \WP_CLI\Utils\get_flag_value( $assoc_args, 'detail', 'short' );
-		$uuid         = '';
-		$ilog         = Log::bootstrap( 'plugin', DECALOG_PRODUCT_SHORTNAME, DECALOG_VERSION );
-		$action       = isset( $args[0] ) ? $args[0] : 'list';
-		$loggers_list = Option::network_get( 'loggers' );
-		if ( isset( $args[1] ) ) {
-			$uuid = $args[1];
-			if ( 'add' === $action ) {
-				$handler_types = new HandlerTypes();
-				$t             = '';
-				foreach ( $handler_types->get_all() as $handler ) {
-					if ( 'system' !== $handler['class'] && strtolower( $uuid ) === strtolower( $handler['id'] ) ) {
-						$t = $uuid;
-					}
-					if ( 'system' === $handler['class'] && strtolower( $uuid ) === strtolower( $handler['id'] ) ) {
-						$t = 'system';
-					}
-				}
-				$uuid = $t;
-			} else {
-				if ( ! array_key_exists( $uuid, $loggers_list ) ) {
-					$uuid = '';
-				} else {
-					$handler_types = new HandlerTypes();
-					foreach ( $handler_types->get_all() as $handler ) {
-						if ( 'system' === $handler['class'] && $loggers_list[$uuid]['handler'] === $handler['id'] ) {
-							$uuid = 'system';
-						}
-					}
-				}
-			}
-		}
-		if ( 'add' === $action && '' === $uuid ) {
-			self::error( 1, $stdout );
-		} elseif ( 'system' === $uuid ) {
-			self::error( 3, $stdout );
-		} elseif ( 'list' !== $action && '' === $uuid ) {
-			self::error( 2, $stdout );
-		}
-		switch ( $action ) {
-			case 'list':
-				$handler_types   = new HandlerTypes();
-				$processor_types = new ProcessorTypes();
-				$loggers         = [];
-				foreach ( $loggers_list as $key => $logger ) {
-					$handler           = $handler_types->get( $logger['handler'] );
-					$logger['type']    = $handler['name'];
-					$logger['uuid']    = $key;
-					$logger['level']   = strtolower( Log::level_name( $logger['level'] ) );
-					$logger['running'] = $logger['running'] ? 'yes' : 'no';
-					$list              = [ 'Standard' ];
-					foreach ( $logger['processors'] as $processor ) {
-						$list[] = $processor_types->get( $processor )['name'];
-					}
-					$logger['processors'] =  implode( ', ', $list );
-					$loggers[]            = $logger;
-				}
-				usort(
-					$loggers,
-					function ( $a, $b ) {
-						return strcmp( strtolower( $a[ 'name' ] ), strtolower( $b[ 'name' ] ) );
-					}
-				);
-				if ( 'full' === $detail ) {
-					$detail = [ 'uuid', 'type', 'name', 'running', 'level', 'processors' ];
-				} else {
-					$detail = [ 'uuid', 'type', 'name', 'running' ];
-				}
-				if ( 'ids' === $format ) {
-					self::write_ids( $loggers, 'uuid' );
-				} else {
-					\WP_CLI\Utils\format_items( $format, $loggers, $detail );
-				}
-				break;
-			case 'start':
-				if ( $loggers_list[$uuid]['running'] ) {
-					self::line( sprintf( 'The logger %s is already running.', $uuid ), $uuid, $stdout );
-				} else {
-					$loggers_list[$uuid]['running'] = true;
-					Option::network_set( 'loggers', $loggers_list );
-					$ilog->info( sprintf( 'Logger "%s" has started.', $loggers_list[ $uuid ]['name'] ) );
-					self::success( sprintf( 'logger %s is now running.', $uuid ), $uuid, $stdout );
-				}
-				break;
-			case 'pause':
-				if ( ! $loggers_list[$uuid]['running'] ) {
-					self::line( sprintf( 'The logger %s is already paused.', $uuid ), $uuid, $stdout );
-				} else {
-					$loggers_list[$uuid]['running'] = false;
-					$ilog->info( sprintf( 'Logger "%s" has been paused.', $loggers_list[ $uuid ]['name'] ) );
-					Option::network_set( 'loggers', $loggers_list );
-					self::success( sprintf( 'logger %s is now paused.', $uuid ), $uuid, $stdout );
-				}
-				break;
-			case 'purge':
-				$loggers_list[$uuid]['uuid'] = $uuid;
-				if ( 'WordpressHandler' !== $loggers_list[$uuid]['handler'] ) {
-					self::warning( sprintf( 'logger %s can\'t be purged.', $uuid ), $uuid, $stdout );
-				} else {
-					\WP_CLI::confirm( sprintf( 'Are you sure you want to purge logger %s?', $uuid ), $assoc_args );
-					$factory = new LoggerFactory();
-					$factory->purge( $loggers_list[$uuid] );
-					$ilog->notice( sprintf( 'Logger "%s" has been purged.', $loggers_list[ $uuid ]['name'] ) );
-					self::success( sprintf( 'logger %s successfully purged.', $uuid ), $uuid, $stdout );
-				}
-				break;
-			case 'clean':
-				$loggers_list[$uuid]['uuid'] = $uuid;
-				if ( 'WordpressHandler' !== $loggers_list[$uuid]['handler'] ) {
-					self::warning( sprintf( 'logger %s can\'t be cleaned.', $uuid ), $uuid, $stdout );
-				} else {
-					$factory = new LoggerFactory();
-					$count   = $factory->clean( $loggers_list[$uuid] );
-					self::log( sprintf( '%d record(s) deleted.', $count ), $stdout );
-					self::success( sprintf( 'logger %s successfully cleaned.', $uuid ), $uuid, $stdout );
-				}
-				break;
-			case 'remove':
-				$loggers_list[$uuid]['uuid'] = $uuid;
-				\WP_CLI::confirm( sprintf( 'Are you sure you want to remove logger %s?', $uuid ), $assoc_args );
-				$factory = new LoggerFactory();
-				$factory->destroy( $loggers_list[$uuid] );
-				$ilog->notice( sprintf( 'Logger "%s" has been removed.', $loggers_list[ $uuid ]['name'] ) );
-				unset( $loggers_list[$uuid] );
-				Option::network_set( 'loggers', $loggers_list );
-				self::success( sprintf( 'logger %s successfully removed.', $uuid ), $uuid, $stdout );
-				break;
-			case 'add':
-				$result = self::logger_add( $uuid, $assoc_args );
-				if ( '' === $result ) {
-					$ilog->error( 'Unable to add a logger.', 1 );
-					self::error( 4, $stdout );
-				} else {
-					$loggers_list = Option::network_get( 'loggers' );
-					$ilog->notice( sprintf( 'Logger "%s" has been saved.', $loggers_list[ $result ]['name'] ) );
-					self::success( sprintf( 'logger %s successfully created.', $result ), $result, $stdout );
-				}
-				break;
-			case 'set':
-				$result = self::logger_modify( $uuid, $assoc_args );
-				if ( '' === $result ) {
-					$ilog->error( 'Unable to modify a logger.', 1 );
-					self::error( 5, $stdout );
-				} else {
-					$loggers_list = Option::network_get( 'loggers' );
-					$ilog->notice( sprintf( 'Logger "%s" has been saved.', $loggers_list[ $result ]['name'] ) );
-					self::success( sprintf( 'logger %s successfully saved.', $result ), $result, $stdout );
-				}
-				break;
-		}
-	}
-
-	/**
-	 * Manage Decalog listener.
-	 *
-	 * ## OPTIONS
-	 *
-	 * <list|enable|disable|auto-on|auto-off>
-	 * : The action to take.
-	 * ---
-	 * default: list
-	 * options:
-	 *  - list
-	 *  - enable
-	 *  - disable
-	 *  - auto-on
-	 *  - auto-off
-	 * ---
-	 *
-	 * [<listener_id>]
-	 * : The id of the listener to perform an action on.
-	 *
-	 * [--detail=<detail>]
-	 * : The details of the output when listing listeners.
-	 * ---
-	 * default: short
-	 * options:
-	 *  - short
-	 *  - full
-	 * ---
-	 *
-	 * [--format=<format>]
-	 * : Allows overriding the output of the command when listing listeners.
-	 * ---
-	 * default: table
-	 * options:
-	 *  - table
-	 *  - json
-	 *  - csv
-	 *  - yaml
-	 *  - ids
-	 *  - count
-	 * ---
-	 *
-	 * [--yes]
-	 * : Answer yes to the confirmation message, if any.
-	 *
-	 * [--stdout]
-	 * : Use clean STDOUT output to use results in scripts. Unnecessary when piping commands because piping is detected by DecaLog.
-	 *
-	 * ## EXAMPLES
-	 *
-	 * Lists configured listeners:
-	 * + wp log listener list
-	 * + wp log listener list --detail=full
-	 * + wp log listener list --format=json
-	 *
-	 * Enables a listener:
-	 * + wp log listener enable wpdb
-	 *
-	 * Disables a listener:
-	 * wp log listener disable wpdb
-	 *
-	 * Activates auto-listening:
-	 * + wp log listener auto-on
-	 * + wp log listener auto-on --yes
-	 *
-	 * Deactivates auto-listening:
-	 * + wp log listener auto-off
-	 * + wp log listener auto-off --yes
-	 *
-	 *
-	 *   === For other examples and recipes, visit https://github.com/Pierre-Lannoy/wp-traffic/blob/master/WP-CLI.md ===
-	 *
-	 */
-	public static function listener( $args, $assoc_args ) {
-		$stdout    = \WP_CLI\Utils\get_flag_value( $assoc_args, 'stdout', false );
-		$format    = \WP_CLI\Utils\get_flag_value( $assoc_args, 'format', 'table' );
-		$detail    = \WP_CLI\Utils\get_flag_value( $assoc_args, 'detail', 'short' );
-		$activated = Option::network_get( 'listeners' );
-		$listeners = [];
-		foreach ( ListenerFactory::$infos as $listener ) {
-			$listener['enabled']        = Option::network_get( 'autolisteners') ? 'auto' : ( in_array( $listener['id'], $activated, true ) ? 'yes' : 'no' );
-			$listener['available']      = $listener['available'] ? 'yes' : 'no';
-			$listeners[$listener['id']] = $listener;
-		}
-		uasort(
-			$listeners,
-			function ( $a, $b ) {
-				return strcmp( strtolower( $a[ 'name' ] ), strtolower( $b[ 'name' ] ) );
-			}
-		);
-		$uuid   = '';
-		$ilog   = Log::bootstrap( 'plugin', DECALOG_PRODUCT_SHORTNAME, DECALOG_VERSION );
-		$action = isset( $args[0] ) ? $args[0] : 'list';
-		if ( isset( $args[1] ) ) {
-			$uuid = strtolower( $args[1] );
-			if ( ! array_key_exists( $uuid, $listeners ) ) {
-				$uuid = '';
-			}
-		}
-		if ( 'list' !== $action && 'auto-on' !== $action && 'auto-off' !== $action && '' === $uuid ) {
-			self::error( 6, $stdout );
-		}
-		switch ( $action ) {
-			case 'list':
-				if ( 'full' === $detail ) {
-					$detail = [ 'id', 'class', 'name', 'product', 'version', 'available', 'enabled' ];
-				} else {
-					$detail = [ 'id', 'name', 'available', 'enabled' ];
-				}
-				if ( 'ids' === $format ) {
-					self::write_ids( $listeners, 'id' );
-				} else {
-					\WP_CLI\Utils\format_items( $format, $listeners, $detail );
-				}
-				break;
-			case 'enable':
-				if ( in_array( $uuid, $activated, true ) ) {
-					self::line( sprintf( 'the listener %s is already enabled.', $uuid ), $uuid, $stdout );
-				} else {
-					$activated[] = $uuid;
-					Option::network_set( 'listeners', $activated );
-					$ilog->info( 'Listeners settings updated.' );
-					self::success( sprintf( 'the listener %s is now enabled.', $uuid ), $uuid, $stdout );
-				}
-				break;
-			case 'disable':
-				if ( ! in_array( $uuid, $activated, true ) ) {
-					self::line( sprintf( 'the listener %s is already disabled.', $uuid ), $uuid, $stdout );
-				} else {
-					$list = [];
-					foreach ( $activated as $listener ) {
-						if ( $listener !== $uuid ) {
-							$list[] = $listener;
-						}
-					}
-					Option::network_set( 'listeners', $list );
-					$ilog->info( 'Listeners settings updated.' );
-					self::success( sprintf( 'the listener %s is now disabled.', $uuid ), $uuid, $stdout );
-				}
-				break;
-			case 'auto-on':
-				if ( Option::network_get( 'autolisteners' ) ) {
-					self::line( 'auto-listening is already activated.', '', $stdout );
-				} else {
-					\WP_CLI::confirm( 'Are you sure you want to activate auto-listening?', $assoc_args );
-					Option::network_set( 'autolisteners', true );
-					$ilog->info( 'Listeners settings updated.' );
-					self::success( 'auto-listening is now activated.', '', $stdout );
-				}
-				break;
-			case 'auto-off':
-				if ( ! Option::network_get( 'autolisteners' ) ) {
-					self::line( 'auto-listening is already deactivated.', '', $stdout );
-				} else {
-					\WP_CLI::confirm( 'Are you sure you want to deactivate auto-listening?', $assoc_args );
-					Option::network_set( 'autolisteners', false );
-					$ilog->info( 'Listeners settings updated.' );
-					self::success( 'auto-listening is now deactivated.', '', $stdout );
-				}
-				break;
-		}
-
 	}
 
 	/**
@@ -1051,43 +423,61 @@ class Wpcli {
 	}
 
 	/**
-	 * Send a message to all running loggers.
+	 * Get information on exit codes.
 	 *
-	 * <info|notice|warning|error|critical|alert>
-	 * : The level of the event.
+	 * ## OPTIONS
 	 *
-	 * <message>
-	 * : The message.
+	 * <list>
+	 * : The action to take.
+	 * ---
+	 * options:
+	 *  - list
+	 * ---
 	 *
-	 * [--code=<code>]
-	 * : The code of the event. Must be a positive integer. Default is 0.
+	 * [--format=<format>]
+	 * : Allows overriding the output of the command when listing exit codes.
+	 * ---
+	 * default: table
+	 * options:
+	 *  - table
+	 *  - json
+	 *  - csv
+	 *  - yaml
+	 *  - ids
+	 *  - count
+	 * ---
 	 *
 	 * ## EXAMPLES
 	 *
-	 * wp log send info 'This is an informational message'
-	 * wp log send warning 'Page not found' --code=404
+	 * Lists available exit codes:
+	 * + wp api exitcode list
+	 * + wp api exitcode list --format=json
 	 *
 	 *
 	 *   === For other examples and recipes, visit https://github.com/Pierre-Lannoy/wp-traffic/blob/master/WP-CLI.md ===
 	 *
 	 */
-	public static function send( $args, $assoc_args ) {
-		$level   = isset( $args[0] ) ? strtolower( $args[0] ) : '';
-		$message = isset( $args[1] ) ? (string) $args[1] : '';
-		$code    = isset( $assoc_args['code'] ) ? (int) $assoc_args['code'] : 0;
-		if ( 0 > $code ) {
-			$code = 0;
+	public static function exitcode( $args, $assoc_args ) {
+		$stdout = \WP_CLI\Utils\get_flag_value( $assoc_args, 'stdout', false );
+		$format = \WP_CLI\Utils\get_flag_value( $assoc_args, 'format', 'table' );
+		$action = isset( $args[0] ) ? $args[0] : 'list';
+		$codes  = [];
+		foreach ( self::$exit_codes as $key => $msg ) {
+			$codes[ $key ] = [ 'code' => $key, 'meaning' => ucfirst( $msg ) ];
 		}
-		if ( ! in_array( $level, ['info', 'notice', 'warning', 'error', 'critical', 'alert' ], true ) ) {
-			\WP_CLI::error( 'forbidden or unknown level.' );
+		switch ( $action ) {
+			case 'list':
+				if ( 'ids' === $format ) {
+					self::write_ids( $codes );
+				} else {
+					\WP_CLI\Utils\format_items( $format, $codes, [ 'code', 'meaning' ] );
+				}
+				break;
 		}
-		$logger = Log::bootstrap( 'core', 'WP-CLI', defined( 'WP_CLI_VERSION' ) ? WP_CLI_VERSION : 'x' );
-		$logger->log( $level, $message, $code );
-		\WP_CLI::success( 'message sent.' );
 	}
 
 	/**
-	 * Get information on exit codes.
+	 * Get information on http status.
 	 *
 	 * ## OPTIONS
 	 *
@@ -1114,19 +504,19 @@ class Wpcli {
 	 * ## EXAMPLES
 	 *
 	 * Lists available exit codes:
-	 * + wp log exitcode list
-	 * + wp log exitcode list --format=json
+	 * + wp api httpstatus list
+	 * + wp api httpstatus list --format=json
 	 *
 	 *
 	 *   === For other examples and recipes, visit https://github.com/Pierre-Lannoy/wp-traffic/blob/master/WP-CLI.md ===
 	 *
 	 */
-	public static function exitcode( $args, $assoc_args ) {
+	public static function httpstatus( $args, $assoc_args ) {
 		$stdout = \WP_CLI\Utils\get_flag_value( $assoc_args, 'stdout', false );
 		$format = \WP_CLI\Utils\get_flag_value( $assoc_args, 'format', 'table' );
 		$action = isset( $args[0] ) ? $args[0] : 'list';
 		$codes  = [];
-		foreach ( self::$exit_codes as $key => $msg ) {
+		foreach ( Http::$http_status_codes as $key => $msg ) {
 			$codes[ $key ] = [ 'code' => $key, 'meaning' => ucfirst( $msg ) ];
 		}
 		switch ( $action ) {
@@ -1266,12 +656,8 @@ add_shortcode( 'traffic-wpcli', [ 'Traffic\Plugin\Feature\Wpcli', 'sc_get_helpfi
 
 if ( defined( 'WP_CLI' ) && WP_CLI ) {
 	\WP_CLI::add_command( 'api tail', [ Wpcli::class, 'tail' ] );
-	/*\WP_CLI::add_command( 'log status', [ Wpcli::class, 'status' ] );
-	\WP_CLI::add_command( 'log logger', [ Wpcli::class, 'logger' ] );
-	\WP_CLI::add_command( 'log type', [ Wpcli::class, 'handler' ] );
-	\WP_CLI::add_command( 'log listener', [ Wpcli::class, 'listener' ] );
-	\WP_CLI::add_command( 'log tail', [ Wpcli::class, 'tail' ] );
-	\WP_CLI::add_command( 'log settings', [ Wpcli::class, 'settings' ] );
-	\WP_CLI::add_command( 'log send', [ Wpcli::class, 'send' ] );
-	\WP_CLI::add_command( 'log exitcode', [ Wpcli::class, 'exitcode' ] );*/
+	\WP_CLI::add_command( 'api status', [ Wpcli::class, 'status' ] );
+	\WP_CLI::add_command( 'api settings', [ Wpcli::class, 'settings' ] );
+	\WP_CLI::add_command( 'api exitcode', [ Wpcli::class, 'exitcode' ] );
+	\WP_CLI::add_command( 'api httpstatus', [ Wpcli::class, 'httpstatus' ] );
 }